Create README.md
Oxyplot is a great open source charting library. But it seems that most of the development is done for .Net and it doesn't provide good examples or even proper documentation for using Oxyplot with WPF.

I struggled a lot while trying to implement a heatmap chart in my WPF project. So, I have decided to create my own repository with sample test code for charting with Oxyplot in WPF. I hope it would be helpful to some other developers.